feat: add jieba-macros for HMM data generation and remove fs operation in build.rs
