+++ src/hmm.rs
@@ -4,6 +4,7 @@ use lazy_static::lazy_static;
 use regex::Regex;
 
 use crate::SplitMatches;
+use jieba_macros::generate_hmm_data;
 
 lazy_static! {
     static ref RE_HAN: Regex = Regex::new(r"([\u{4E00}-\u{9FD5}]+)").unwrap();
@@ -26,9 +27,9 @@ pub type StateSet = [f64; NUM_STATES];
 /// to that state.
 ///
 /// WARNING: The data file format for hmm.model comments imply one can
-/// reassign the index values of each state at the top but `build.rs`
+/// reassign the index values of each state at the top but `jieba-macros`
 /// currently ignores the mapping. Do not reassign these indicies without
-/// verifying how it interacts with `build.rs`.  These indicies must also
+/// verifying how it interacts with `jieba-macros`.  These indicies must also
 /// match the order if ALLOWED_PREV_STATUS.
 #[derive(Debug, PartialEq, Eq, Hash, PartialOrd, Ord, Clone, Copy)]
 pub enum State {
@@ -52,7 +53,7 @@ static ALLOWED_PREV_STATUS: [[State; 2]; NUM_STATES] = [
     [State::Single, State::End],
 ];
 
-include!(concat!(env!("OUT_DIR"), "/hmm_prob.rs"));
+generate_hmm_data!();
 
 const MIN_FLOAT: f64 = -3.14e100;
